This is for using the rule of sines
I was scrolling in a Discord server that is concerned with school work. I saw someone who needed help with a question that was related to The Rule of Sines. I helped him and to make sure that my answer was correct, I used an online Sine Rule calculator. So, I decided to make my own Sine Rule calculator. Btw it's not finished yet, this is only for one situation, which is when the question gives 2 angles and 1 side. Also, the code isn't perfect yet, even for the 1st situation. I'm gonna continue tomorrow. Today was just to recover some C concepts lol.